+++ CHANGES.rst
@@ -6,8 +6,9 @@ Changelog
 
 - Added dmsoutgoing_email type
   [ndemonte]
-
 - Added dmsincoming_email type
+  [sgeulette]
+- Added collective.ckeditortemplates
   [sgeulette]
 
 2.3 (2020-10-08)